fix: remove offline training as it blocks the ci workflow
