+++ tests/integration/faiss_psql/deprecated_test_train_offline.py
@@ -4,8 +4,6 @@ import numpy as np
 import pytest
 from jina import Document, DocumentArray, Flow
 
-from jinahub.indexers.compound.FaissPostgresIndexer import FaissPostgresIndexer
-
 cur_dir = os.path.dirname(os.path.abspath(__file__))
 compose_yml = os.path.join(cur_dir, 'docker-compose.yml')
