+++ tests/integration/faiss_psql/test_train_online.py
@@ -35,3 +35,63 @@ def test_online_train(docker_compose, index_key):
 
         status = f.post(on='/status', return_results=True)[0].docs[0].tags
         assert int(status['active_docs']) == 1024
+
+
+@pytest.mark.parametrize('docker_compose', [compose_yml], indirect=['docker_compose'])
+def test_online_train_2_flows(docker_compose):
+    docs = get_documents(1024)
+    with Flow().add(
+        uses='FaissPostgresIndexer', uses_with={'index_key': 'IVF64,PQ32'}
+    ) as f:
+        f.post(on='/index', inputs=docs)
+
+        with Flow().add(
+            uses='FaissPostgresIndexer', uses_with={'index_key': 'IVF64,PQ32'}
+        ) as train_f:
+            train_f.post(on='/train')
+
+        f.post(on='/sync')
+
+        result = f.post(on='/search', inputs=get_documents(10), return_results=True)[0]
+        for doc in result.docs:
+            assert len(doc.matches) == 10
+
+        f.post(on='/index', inputs=get_documents(10, index_start=1024))
+
+        with Flow().add(
+            uses='FaissPostgresIndexer', uses_with={'index_key': 'IVF64,PQ32'}
+        ) as train_f:
+            train_f.post(on='/train', parameters={'force': True})
+
+        f.post(on='/sync')
+        result = f.post(on='/search', inputs=get_documents(10), return_results=True)[0]
+        for doc in result.docs:
+            assert len(doc.matches) == 10
+
+        status = f.post(on='/status', return_results=True)[0].docs[0].tags
+        assert int(status['active_docs']) == 1034
+
+
+@pytest.mark.parametrize('docker_compose', [compose_yml], indirect=['docker_compose'])
+def test_online_train_3_flows(docker_compose):
+    docs = get_documents(1024)
+    with Flow().add(
+        uses='FaissPostgresIndexer', uses_with={'index_key': 'IVF64,PQ32'}
+    ) as f:
+        f.post(on='/index', inputs=docs)
+
+    with Flow().add(
+        uses='FaissPostgresIndexer', uses_with={'index_key': 'IVF64,PQ32'}
+    ) as f:
+        f.post(on='/train')
+
+    with Flow().add(
+        uses='FaissPostgresIndexer', uses_with={'index_key': 'IVF64,PQ32'}
+    ) as f:
+        f.post(on='/sync')
+        result = f.post(on='/search', inputs=get_documents(10), return_results=True)[0]
+        for doc in result.docs:
+            assert len(doc.matches) == 10
+
+        status = f.post(on='/status', return_results=True)[0].docs[0].tags
+        assert int(status['active_docs']) == 1024
